+++ halo/evmstaking2/keeper/helpers.go
@@ -1,0 +1,28 @@
+package keeper
+
+import (
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+)
+
+// mustGetABI returns the metadata's ABI as an abi.ABI type.
+// It panics on error.
+func mustGetABI(metadata *bind.MetaData) *abi.ABI {
+	abi, err := metadata.GetAbi()
+	if err != nil {
+		panic(err)
+	}
+
+	return abi
+}
+
+// mustGetEvent returns the event with the given name from the ABI.
+// It panics if the event is not found.
+func mustGetEvent(abi *abi.ABI, name string) abi.Event {
+	event, ok := abi.Events[name]
+	if !ok {
+		panic("event not found")
+	}
+
+	return event
+}
